Prevent the iOS touchstart fix to break on Android
iOS (6, 7.0, 7.1) triggers another touchstart with the same identifier
after an alert(), but the Touch identifier stays the same between these
two touchstart events, so we can use it to prevent another triggering of
the action.

On iOS, this identifier is a number that changes on every new Touch
event, except for this particular case.

On Android, this identifier starts with 0 and is incremented for every
Touch object, so we can have the same identifier for two different
events, which means that we have to detect the uniqueness of the
identifier before relying on it: if the identifier is greater than
IDENTIFIER_UNIQUENESS_LIMIT, it means that the number is unique.